the big come down

+++ bash/run.sh
@@ -1,9 +1,17 @@
 #!/usr/bin/env bash
 
-RUSTFLAGS="-Awarnings" cargo run --bin server &
+RUSTFLAGS="-Awarnings" cargo run --bin nimbus-server &
 server_process=$!
 
-RUSTFLAGS="-Awarnings" cargo run --bin client-desktop
+sleep 1
+
+if ! kill -0 "$server_process" 2>/dev/null; then
+    wait "$server_process"
+    exit_code=$?
+    exit "$exit_code"
+fi
+
+RUSTFLAGS="-Awarnings" cargo run --bin nimbus-desktop
 
 kill "$server_process" 2>/dev/null
 wait "$server_process" 2>/dev/null
